requirements.txt
requests
bs4
logging
csv
psycopg2
openai-whisper
